Reimplement scaling logic
I had some initial misunderstandings for how the x/y translation and
scaling was supposed to work which resulted in some added complexity to
handle issues. For example, the original implementations all swapped x
and y but then re-swapped x and y in the driver to account. Now all the
swapping is removed.