Embedd relevant portions of robotgo
This patch removes all but the basic system dependencies by embedding a
reduced version of the robotgo. This should enable more portable
binaries by limiting what a system needs to have installed to only
pre-installed components like x11 or the OSX system libraries.

This may create a long term support issues as we'll need to rebuild the
embedded package from the original source after each major change or
when a new supported system is added.